Add edit-layer-mode-form as generic layer component.
This is a generic component that will be used to edit
layer with mode features, feature-metric and cluster.
Actually this is just the same content as
edit-layer-features.
