+++ src/app/modules/map-config/components/edit-layer-feature-metric/edit-layer-feature-metric.component.html
@@ -1,4 +1,4 @@
-<app-edit-layer-features [submit]="submitSubject.asObservable()">
+<app-edit-layer-mode-form [submit]="submitSubject.asObservable()">
     <div additional-geometry>
         <app-config-element>
             <div form-fields>
@@ -15,4 +15,4 @@
             </div>
         </app-config-element>
     </div>
-</app-edit-layer-features>
+</app-edit-layer-mode-form>
