+++ src/app/modules/map-config/components/edit-layer-feature-metric/edit-layer-feature-metric.component.ts
@@ -48,6 +48,9 @@ export class EditLayerFeatureMetricComponent extends EditLayerFeatureMetricCompo
   }
 
   public ngOnInit() {
+
+    super.ngOnInit();
+
     // by getting a reference to the embedded form in this variable,
     // it will used by the parent ControlValueAccessor implementation to write values on-the-fly
     this.formFg = this.embeddedFeaturesComponent.formFg;
