+++ src/app/modules/map-config/components/edit-layer-features/edit-layer-features.component.spec.ts
@@ -1,35 +1,17 @@
-import { Spectator, createComponentFactory, mockProvider } from '@ngneat/spectator';
 import { EditLayerFeaturesComponent } from './edit-layer-features.component';
-import { MockComponent, MockDirective } from 'ng-mocks';
-import { Subject } from 'rxjs';
-import { DefaultValuesService } from '@services/default-values/default-values.service';
-import { HttpClient } from '@angular/common/http';
+import { Spectator, createComponentFactory } from '@ngneat/spectator';
+import { EditLayerModeFormComponent } from '../edit-layer-mode-form/edit-layer-mode-form.component';
 import { ConfigElementComponent } from '@shared-components/config-element/config-element.component';
-import { AlertOnChangeDirective } from '@shared-directives/alert-on-change/alert-on-change.directive';
-import { ResetOnChangeDirective } from '@shared-directives/reset-on-change/reset-on-change.directive';
-import { ColorPickerWrapperComponent } from '@shared-components/color-picker-wrapper/color-picker-wrapper.component';
-import { CollectionService } from '@services/collection-service/collection.service';
-import { ArlasColorGeneratorLoader } from 'arlas-wui-toolkit';
-import { MainFormService } from '@services/main-form/main-form.service';
-import { PropertySelectorComponent } from '@shared-components/property-selector/property-selector.component';
+import { Subject } from 'rxjs';
+import { MockComponent } from 'ng-mocks';
 
 describe('EditLayerFeaturesComponent', () => {
   let spectator: Spectator<EditLayerFeaturesComponent>;
   const createComponent = createComponentFactory({
     component: EditLayerFeaturesComponent,
     declarations: [
-      MockComponent(ConfigElementComponent),
-      MockDirective(AlertOnChangeDirective),
-      MockDirective(ResetOnChangeDirective),
-      MockComponent(ColorPickerWrapperComponent),
-      MockComponent(PropertySelectorComponent),
-    ],
-    mocks: [
-      HttpClient,
-      CollectionService,
-      ArlasColorGeneratorLoader,
-      MainFormService,
-      DefaultValuesService
+      MockComponent(EditLayerModeFormComponent),
+      MockComponent(ConfigElementComponent)
     ]
   });
 
@@ -43,7 +25,4 @@ describe('EditLayerFeaturesComponent', () => {
     expect(spectator.component).toBeTruthy();
   });
 
-  it('should contain 4 steps', () => {
-    expect(spectator.queryAll('mat-step-header')).toHaveLength(4);
-  });
 });
