+++ src/app/modules/map-config/components/edit-layer-features/edit-layer-features.component.ts
@@ -16,21 +16,10 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 */
-import { STEPPER_GLOBAL_OPTIONS } from '@angular/cdk/stepper';
-import { Component, forwardRef, OnInit, ViewChild } from '@angular/core';
-import {
-  FormBuilder, NG_VALIDATORS, NG_VALUE_ACCESSOR, FormControl
-} from '@angular/forms';
-import { MatSliderChange } from '@angular/material/slider';
-import { MatStepper } from '@angular/material/stepper';
-import { CollectionService, FIELD_TYPES } from '@services/collection-service/collection.service';
-import { FormBuilderWithDefaultService } from '@services/form-builder-with-default/form-builder-with-default.service';
-import { MainFormService } from '@services/main-form/main-form.service';
-import { NGXLogger } from 'ngx-logger';
+import { Component, OnInit, forwardRef } from '@angular/core';
 import { EditLayerFeaturesComponentForm } from './edit-layer-features.component.form';
-import { GEOMETRY_TYPE } from './models';
-import { PROPERTY_SELECTOR_SOURCE, PROPERTY_TYPE } from '@shared-components/property-selector/models';
-import { ensureMinLessThanMax } from '@utils/tools';
+import { NG_VALUE_ACCESSOR, NG_VALIDATORS } from '@angular/forms';
+import { NGXLogger } from 'ngx-logger';
 
 @Component({
   selector: 'app-edit-layer-features',
@@ -46,118 +35,21 @@ import { ensureMinLessThanMax } from '@utils/tools';
       provide: NG_VALIDATORS,
       useExisting: forwardRef(() => EditLayerFeaturesComponent),
       multi: true
-    },
-    {
-      provide: STEPPER_GLOBAL_OPTIONS,
-      useValue: { showError: true }
     }
   ]
 })
-// ControlValueAccessor: see https://christianlydemann.com/form-validation-with-controlvalueaccessor/
 export class EditLayerFeaturesComponent extends EditLayerFeaturesComponentForm implements OnInit {
 
-  @ViewChild('stepper', { static: false }) stepper: MatStepper;
-
-  public ensureMinLessThanMax = ensureMinLessThanMax;
-
-  public GEOMETRY_TYPE = GEOMETRY_TYPE;
-  public PROPERTY_TYPE = PROPERTY_TYPE;
-  public colorFgSources = Object.values(PROPERTY_SELECTOR_SOURCE);
-  public widthFgSources = [PROPERTY_SELECTOR_SOURCE.fix, PROPERTY_SELECTOR_SOURCE.interpolated];
-  public radiusFgSources = [PROPERTY_SELECTOR_SOURCE.fix, PROPERTY_SELECTOR_SOURCE.interpolated];
-  public collectionGeoFields: string[] = [];
-  public collectionKeywordFields: string[] = [];
-  public collectionIntegerFields: string[] = [];
-
   constructor(
-    protected formBuilderDefault: FormBuilderWithDefaultService,
-    protected formBuilder: FormBuilder,
-    protected logger: NGXLogger,
-    public mainformService: MainFormService,
-    public collectionService: CollectionService
-  ) {
-    super(formBuilderDefault, formBuilder, logger);
+    protected logger: NGXLogger) {
+    super(logger);
   }
 
   ngOnInit() {
     super.ngOnInit();
-    this.initCollectionRelatedFields();
-    this.initEnableWidthOrRadiusFg();
+    // by getting a reference to the embedded form in this variable,
+    // it will used by the parent ControlValueAccessor implementation to write values on-the-fly
+    this.formFg = this.embeddedFeaturesComponent.formFg;
   }
-
-  protected onSubmit() {
-    super.onSubmit();
-    // activate stepper validation on submit
-    this.stepper.steps.setDirty();
-    this.stepper.steps.forEach(s => s.interacted = true);
-  }
-
-  private initCollectionRelatedFields() {
-    this.collectionCtrl.valueChanges.subscribe(c => {
-      if (!c) {
-        return;
-      }
-      this.collectionService.getCollectionFields(c, [FIELD_TYPES.GEOPOINT, FIELD_TYPES.GEOSHAPE])
-        .subscribe(
-          f => this.collectionGeoFields = f);
-      this.collectionService.getCollectionFields(c, [FIELD_TYPES.KEYWORD])
-        .subscribe(
-          f => this.collectionKeywordFields = f);
-      this.collectionService.getCollectionFields(c, [FIELD_TYPES.LONG, FIELD_TYPES.INTEGER, FIELD_TYPES.DATE])
-        .subscribe(
-          f => this.collectionIntegerFields = f);
-    });
-  }
-
-  /**
-   * widthFg and radiusFg are conditionally displayed, once they have been displayed, their subform has been
-   * registred into the main form and their validation works even if they aren't displayed anymore.
-   * The solution is to enable only the expected form group.
-   */
-  private initEnableWidthOrRadiusFg() {
-    this.geometryTypeCtrl.valueChanges.subscribe(v => {
-      const geoEnableDisable = [{
-        geometry: GEOMETRY_TYPE.line,
-        enabled: [this.widthFg],
-        disabled: [this.radiusFg]
-      },
-      {
-        geometry: GEOMETRY_TYPE.circle,
-        enabled: [this.radiusFg],
-        disabled: [this.widthFg]
-      },
-      {
-        geometry: GEOMETRY_TYPE.fill,
-        enabled: [this.radiusFg, this.widthFg],
-        disabled: []
-      }].find(elmt => elmt.geometry === v);
-
-      if (!!geoEnableDisable) {
-        geoEnableDisable.enabled.forEach(c => c.enable());
-        geoEnableDisable.disabled.forEach(c => c.disable());
-      }
-    });
-    this.geometryTypeCtrl.updateValueAndValidity({ onlySelf: true, emitEvent: true });
-  }
-
-  writeValue(obj: any): void {
-    super.writeValue(obj);
-    if (obj) {
-      this.collectionCtrl.updateValueAndValidity({ onlySelf: true, emitEvent: true });
-    }
-  }
-
-  public checkZoom(event: MatSliderChange, source: string) {
-    if (source === 'min') {
-      if (event.value > this.zoomMaxCtrl.value) {
-        this.zoomMaxCtrl.setValue(event.value);
-      }
-    } else if (source === 'max') {
-      if (event.value < this.zoomMinCtrl.value) {
-        this.zoomMinCtrl.setValue(event.value);
-      }
-    }
-  }
-
 
 }
