+++ src/app/modules/map-config/components/edit-layer-mode-form/edit-layer-mode-form.component.scss
@@ -16,24 +16,3 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 */
-import { FormGroup, Validators } from '@angular/forms';
-import { FormBuilderWithDefaultService } from '@services/form-builder-with-default/form-builder-with-default.service';
-
-export abstract class EditLayerComponentForm {
-    public layerFg: FormGroup;
-
-    constructor(
-        protected formBuilderDefault: FormBuilderWithDefaultService
-    ) {
-        this.layerFg = this.formBuilderDefault.group('map.layer', {
-            name: ['', Validators.required],
-            mode: ['', Validators.required],
-            id: [''],
-            modeFg: ['', Validators.required]
-        });
-    }
-
-    get modeFg() {
-        return this.layerFg.get('modeFg');
-    }
-}
