+++ src/app/modules/map-config/components/edit-layer/edit-layer.component.spec.ts
@@ -1,15 +1,17 @@
 import { EditLayerComponent } from './edit-layer.component';
 import { Spectator, createComponentFactory, mockProvider } from '@ngneat/spectator';
-import { MockComponent } from 'ng-mocks';
+import { MockComponent, MockDirective } from 'ng-mocks';
 import { Component } from '@angular/core';
 import { RouterTestingModule } from '@angular/router/testing';
 import { ConfigElementComponent } from '@shared-components/config-element/config-element.component';
-import { EditLayerFeaturesComponent } from '../edit-layer-features/edit-layer-features.component';
+import { EditLayerModeFormComponent } from '../edit-layer-mode-form/edit-layer-mode-form.component';
 import { MainFormService } from '@services/main-form/main-form.service';
 import { FormBuilderWithDefaultService } from '@services/form-builder-with-default/form-builder-with-default.service';
 import { FormArray } from '@angular/forms';
 import { EditLayerFeatureMetricComponent } from '../edit-layer-feature-metric/edit-layer-feature-metric.component';
 import { ObjectvaluesPipe } from '@shared/pipes/objectvalues.pipe';
+import { EditLayerFeaturesComponent } from '../edit-layer-features/edit-layer-features.component';
+import { ResetOnChangeDirective } from '@shared-directives/reset-on-change/reset-on-change.directive';
 
 @Component({ template: '' }) class DummyComponent { }
 
@@ -32,6 +34,7 @@ describe('EditLayerComponent', () => {
       MockComponent(ConfigElementComponent),
       MockComponent(EditLayerFeaturesComponent),
       MockComponent(EditLayerFeatureMetricComponent),
+      MockDirective(ResetOnChangeDirective),
       ObjectvaluesPipe
     ]
   });
